Make for commoncrypto and some basic playgrounds
