+++ Makefile
@@ -13,6 +13,13 @@ imports = \
 xcodeproj:
 	swift package generate-xcodeproj
 
+bootstrap: common-crypto-mm xcodeproj
+
+common-crypto-mm:
+	-@mkdir "$(COMMON_CRYPTO_PATH)" || true
+	-@touch "$(COMMON_CRYPTO_MODULE_PATH)" || true
+	-@echo "$$COMMON_CRYPTO_MODULE_MAP" > "$(COMMON_CRYPTO_MODULE_PATH)"
+
 linux-main:
 	sourcery \
 		--sources ./Tests/ \
@@ -43,3 +50,16 @@ test-swift:
 	swift test
 
 test-all: test-linux test-mac test-ios
+
+SDK_PATH = $(shell xcrun --show-sdk-path)
+FRAMEWORKS_PATH = $(SDK_PATH)/System/Library/Frameworks
+COMMON_CRYPTO_PATH = $(FRAMEWORKS_PATH)/CommonCrypto.framework
+COMMON_CRYPTO_MODULE_PATH = $(COMMON_CRYPTO_PATH)/module.map
+define COMMON_CRYPTO_MODULE_MAP
+module CommonCrypto [system] {
+  header "$(SDK_PATH)/usr/include/CommonCrypto/CommonCrypto.h"
+  header "$(SDK_PATH)/usr/include/CommonCrypto/CommonRandom.h"
+  export *
+}
+endef
+export COMMON_CRYPTO_MODULE_MAP
