+++ Web.playground/Pages/Css.xcplaygroundpage/Contents.swift
@@ -1,0 +1,9 @@
+import Css
+import Prelude
+
+let css = body % (
+  padding(all: .rem(2))
+    <> background(Color.hsl(60, 0.5, 0.8))
+)
+
+print(render(css: css))
